Gave FileLogger Its own file
Changed ReadMe
